restructered everything
