+++ .env
@@ -1,4 +1,4 @@
-                "MONGODB_USERNAME": "dev",
-                "MONGODB_PASSWORD": "90NUVASWdA6dBxuz",
-                "MONGODB_ENDPOINT": "localhost:27017",
-                "MONGODB_DATABASE": "akira_dev"
+"MONGODB_USERNAME": "dev",
+"MONGODB_PASSWORD": "90NUVASWdA6dBxuz",
+"MONGODB_ENDPOINT": "localhost:27017",
+"MONGODB_DATABASE": "akira_dev"
