+++ .gitignore
@@ -16,3 +16,5 @@
 
 .vscode/
 _debug_bin
+
+.env
