+++ main.go
@@ -5,6 +5,8 @@ import (
 	"github.com/gin-contrib/sessions"
 	"github.com/gin-contrib/sessions/cookie"
 	"github.com/gin-gonic/gin"
+	"github.com/polaceka/Akira-API/middleware"
+	"github.com/polaceka/Akira-API/routes"
 )
 
 func main() {
@@ -19,31 +21,31 @@ func main() {
 
 	r.Use(cors.New(config))
 
-	r.POST("/login", login)
-	r.GET("/logout", logout)
-	r.POST("/gen", pwgen)
+	r.POST("/login", routes.Login)
+	r.GET("/logout", routes.Logout)
+	r.POST("/gen", routes.Pwgen)
 
 	// Routing API V1
 	v1 := r.Group("/v1")
-	v1.Use(authRequired)
+	v1.Use(middleware.AuthRequired)
 	{
 		// Tacks
-		v1.GET("/track", handlerGetTracks)
-		v1.GET("/track/:id", handlerGetOneTrack)
-		v1.POST("/track", handlerCreateTrack)
+		v1.GET("/track", routes.GetTracks)
+		v1.GET("/track/:id", routes.GetOneTrack)
+		v1.POST("/track", routes.CreateTrack)
 
 		// Events
-		v1.GET("/event", handlerGetEvents)
-		v1.GET("/event/:id", handlerGetOneEvent)
-		v1.POST("/event", handlerCreateEvent)
+		v1.GET("/event", routes.GetEvents)
+		v1.GET("/event/:id", routes.GetOneEvent)
+		v1.POST("/event", routes.CreateEvent)
 		/*
 			v1.Get("/event/:name/*groupe") ?
 		*/
 	}
 
 	// currently for dev ussage
-	r.GET("/me", authRequired, me)
-	r.GET("/status", authRequired, status)
+	r.GET("/me", middleware.AuthRequired, routes.Me)
+	r.GET("/status", middleware.AuthRequired, routes.Status)
 
 	r.Run(":8080") // listen and serve on 0.0.0.0:8080
 }
